+++ packages/andromeda-math/src/lib.rs
@@ -3,5 +3,6 @@ pub mod curve;
 pub mod date_time;
 pub mod distance;
 pub mod graph;
+pub mod matrix;
 pub mod point;
 pub mod shunting;